requirements.txt
# Core dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM and AI
langchain>=0.1.0
langchain-community>=0.0.20
ollama>=0.1.0

# AWS services
boto3>=1.34.0

# Vector database
qdrant-client>=1.7.0

# CLI and UI
click>=8.1.7
rich>=13.7.0

# Async and utilities
aiohttp>=3.9.0
pyyaml>=6.0.1

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.0
mypy>=1.7.0
